More graphs added

+++ src/com/favor/util/Algorithms.java
@@ -61,7 +61,7 @@ public class Algorithms {
    * @param untilDate
    * @return
    */
-  public static double charRatio (String address, long fromDate, long untilDate) {
+  public static long charRatio (String address, long fromDate, long untilDate) {
 	  
 	  //calls character count
 	  long [] values= charCount(address, fromDate, untilDate);
@@ -69,7 +69,7 @@ public class Algorithms {
 	  Debug.log(values[0] + "");
 	  
 	  //some kewl casting here jk
-	  double ratio = values[1]/(float)values[0];
+	  long ratio = (long) (values[1]/(float)values[0]);
 	  return ratio;
   }
